+++ ons_alpha/bundles/forms.py
@@ -1,6 +1,4 @@
-from django import forms
 from django.core.exceptions import ValidationError
-from django.utils.translation import gettext as _
 from wagtail.admin.forms import WagtailAdminModelForm
 
 
@@ -22,21 +20,3 @@ class BundleAdminForm(WagtailAdminModelForm):
                     raise ValidationError(f"{page} is already in an active bundle ({page.active_bundle})")
 
         return cleaned_data
-
-
-class AddToBundleForm(forms.Form):
-    def __init__(self, *args, **kwargs):
-        # imported inline to avoid circular imports
-        from .models import Bundle
-        from .viewsets import BundleChooserWidget
-
-        self.page_to_add = kwargs.pop("page_to_add")
-
-        super().__init__(*args, **kwargs)
-
-        self.fields["bundle"] = forms.ModelChoiceField(
-            queryset=Bundle.active_objects.all(),
-            widget=BundleChooserWidget(),
-            label=_("Bundle"),
-            help_text=_("Select a bundle for this page."),
-        )
